requirements.txt
Flask
python-swiftclient
python-keystoneclient
pyDes